Code Refactoring - broken down to components
